Add a basic shell script and instructions
This is obviously not a real operator yet, but it's a start of how
things can be done. The real operator will need to import code from
code-ready/crc to do the actual reconciling, with a fork likely needed
to handle reconciling on a cluster instead of on a laptop.