Add a new 'admin' user with a random password
There is now an 'admin' user with a random password. The 'developer'
user no longer exists. The kubeadmin password and client cert still
work, but the next step is removing those so that there are no common
credentials across any of these instances.

Once all that works with the 'admin' user, we can try to get the
username back to 'kubeadmin' (it conflicts w/ the hardcoded kubeadmin
today) and look at adding a 'developer' user back with a random password.