Add OCP 4.5.0-rc.2 bundle, simplify bundle mgmt
This adds an OCP 4.5.0-rc.2 bundle and simplifies the management of
bundles a bit. Things are still pretty hardcoded, but at least not
just splatted into the controller code with if statements for each
bundle.

The new `spec.bundleName` is now the preferred way to specify a
bundle, with valid options being `ocp445`, `ocp450rc1`, and
`ocp450rc2`.