Speed up initial persistent cluster creation with HTTP images
There's now an optional `spec.url` on `CrcBundle` resources that can
point to a HTTP/HTTPS URL of this VM image. For non-persistent
clusters, the `spec.image` is always used. For persistent clusters,
`spec.url` will be used if given because that avoids needing to use
temporary scratch space for VM image extraction when creating the PVC
for the cluster.

In my simple testing, this shaves approximately 7 minutes off the
initial cluster creation time for persistent clusters.